pyproject.toml
[project]
name = "desafio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "mcp>=1.27.0",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.2.0",
    "pydantic>=2.13.3",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "uvicorn>=0.44.0",
]
